# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Management System API
  description: |
    API для системы управления рестораном с поддержкой заказов, меню, блюд и управления сотрудниками.

    Система поддерживает раздельную авторизацию для сотрудников и клиентов через JWT токены.

    Роли пользователей:
    - **Admin**: Полный доступ ко всем операциям
    - **Manager**: Управление меню, блюдами, заказами, назначение курьеров
    - **Courier**: Просмотр назначенных заказов, изменение статуса доставки
    - **Client**: Просмотр меню, создание заказов, управление своими данными
  version: 1.0.0
  contact:
    name: API Support
    email: support@restaurant.com
...
components:
  securitySchemes:
    jwtEmployee:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для сотрудников (Admin, Manager, Courier)
    jwtClient:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для клиентов
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, exception]
      properties:
        error:
          type: string
          example: "Internal Server Error"
        exception:
          type: string
          example: "An unexpected error occurred while processing the request"
        message:
          type: string
          example: "Invalid credentials"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation Error"
        message:
          type: string
          example: "Invalid input data"
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            login: ["Login must be between 4 and 16 characters"]
            password: ["Password is required"]
...
    Dish:
      type: object
      required: [id, name, weight, calories, price]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Стейк из говядины"
        weight:
          type: number
          format: float
          example: 300.5
        calories:
          type: integer
          example: 450
        price:
          type: number
          format: decimal
          example: 1200.00
        description:
          type: string
          example: "Сочный стейк из мраморной говядины"
    DishCreate:
      type: object
      required: [name, weight, calories, price]
      properties:
        name:
          type: string
          example: "Стейк из говядины"
        weight:
          type: number
          format: float
          example: 300.5
        calories:
          type: integer
          example: 450
        price:
          type: number
          format: decimal
          example: 1200.00
        description:
          type: string
          example: "Сочный стейк из мраморной говядины"
    DishUpdate:
      type: object
      properties:
        name:
          type: string
          example: "Стейк из говядины"
        weight:
          type: number
          format: float
          example: 300.5
        calories:
          type: integer
          example: 450
        price:
          type: number
          format: decimal
          example: 1200.00
        description:
          type: string
          example: "Сочный стейк из мраморной говядины"
...
    Ingredient:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Говядина"
...
    IngredientUpdate:
      type: object
      properties:
        name:
          type: string
          example: "Говядина"
    DishIngredient:
      type: object
      properties:
        dish_id:
          type: integer
          example: 1
        ingredient_id:
          type: integer
          example: 1
        dish:
          $ref: '#/components/schemas/Dish'
        ingredient:
          $ref: '#/components/schemas/Ingredient'
    DishIngredientCreate:
      type: object
      required: [ingredient_id]
      properties:
        ingredient_id:
          type: integer
          example: 1
...
    DishesList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Dish'
        total:
          type: integer
          example: 30
...
paths:
...
  /dishes:
    get:
      tags: [Dishes]
      summary: Получить список блюд
      description: "Доступ: Все авторизованные"
      operationId: getDishes
      security:
        - jwtEmployee: []
        - jwtClient: []
      responses:
        '200':
          description: Список блюд
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishesList'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Dishes]
      summary: Создать блюдо
      description: "Доступ: Admin, Manager"
      operationId: createDish
      security:
        - jwtEmployee: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishCreate'
      responses:
        '201':
          description: Блюдо создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dishes/{id}:
    get:
      tags: [Dishes]
      summary: Получить блюдо по ID
      description: "Доступ: Все авторизованные"
      operationId: getDishById
      security:
        - jwtEmployee: []
        - jwtClient: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Блюдо
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Dishes]
      summary: Полное обновление блюда
      description: "Доступ: Admin, Manager"
      operationId: updateDish
      security:
        - jwtEmployee: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishUpdate'
      responses:
        '200':
          description: Блюдо обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Dishes]
      summary: Частичное обновление блюда
      description: "Доступ: Admin, Manager"
      operationId: patchDish
      security:
        - jwtEmployee: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishUpdate'
      responses:
        '200':
          description: Блюдо обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Dishes]
      summary: Удалить блюдо
      description: "Доступ: Admin, Manager"
      operationId: deleteDish
      security:
        - jwtEmployee: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Блюдо удалено
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /dishes/{dishId}/ingredients:
    get:
      tags: [Dishes]
      summary: Получить ингредиенты блюда
      description: "Доступ: Все авторизованные"
      operationId: getDishIngredients
      security:
        - jwtEmployee: []
        - jwtClient: []
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список ингредиентов
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ingredient'
                  total:
                    type: integer
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Блюдо не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Dishes]
      summary: Добавить ингредиент к блюду
      description: "Доступ: Admin, Manager"
      operationId: createDishIngredient
      security:
        - jwtEmployee: []
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishIngredientCreate'
      responses:
        '201':
          description: Ингредиент добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DishIngredient'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Блюдо или ингредиент не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dishes/{dishId}/ingredients/{ingredientId}:
    delete:
      tags: [Dishes]
      summary: Удалить ингредиент из блюда
      description: "Доступ: Admin, Manager"
      operationId: deleteDishIngredient
      security:
        - jwtEmployee: []
      parameters:
        - name: dishId
          in: path
          required: true
          schema:
            type: integer
        - name: ingredientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Ингредиент удален
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Связь не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /ingredients/{id}:
    get:
      tags: [Ingredients]
      summary: Получить ингредиент по ID
      description: "Доступ: Все авторизованные"
      operationId: getIngredientById
      security:
        - jwtEmployee: []
        - jwtClient: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ингредиент
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Ingredients]
      summary: Полное обновление ингредиента
      description: "Доступ: Admin, Manager"
      operationId: updateIngredient
      security:
        - jwtEmployee: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientUpdate'
      responses:
        '200':
          description: Ингредиент обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Ingredients]
      summary: Частичное обновление ингредиента
      description: "Доступ: Admin, Manager"
      operationId: patchIngredient
      security:
        - jwtEmployee: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientUpdate'
      responses:
        '200':
          description: Ингредиент обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Ingredients]
      summary: Удалить ингредиент
      description: "Доступ: Admin, Manager"
      operationId: deleteIngredient
      security:
        - jwtEmployee: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Ингредиент удален
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...